fix pathes in deployment script

+++ postdeploy.sh
@@ -3,9 +3,12 @@
 updatedFiles=$(git log --format="%H" -n 2 | xargs git diff --name-only)
 for file in $updatedFiles
 do
-  rsync -aR $file $DEPLOYPATH/$file
+  if [ -f "./$file" ]
+  then
+    rsync -aR $file $DEPLOYPATH/$file
+  fi
 done
 
 cd ~/repositories/qudsinfo-wp-theme/layout
-~/bin/npm run build
+~/bin/npm run css:build
 rsync -aR dist $DEPLOYPATH/layout/dist
